# refactor(agora): modernize Agora and Model-AD frontend by migrating the bundler to Vite (AG-1791)