--- a/apps/agora/app/Dockerfile
+++ b/apps/agora/app/Dockerfile
@@ -1,6 +1,6 @@
 FROM mirror.gcr.io/node:20.7.0-alpine
 
-ENV APP_DIR=/app
+ENV APP_DIR=/opt/app
 
 RUN apk add --no-cache curl envsubst jq su-exec
 
@@ -10,9 +10,8 @@ COPY apps/agora/app/docker-entrypoint.d /docker-entrypoint.d
 RUN chmod +x docker-entrypoint.sh /docker-entrypoint.d/*
 
 WORKDIR ${APP_DIR}
-COPY dist/apps/agora/app/browser/server ${APP_DIR}
-# The path of the destination folder must be the same as the path specified in server.ts.
-COPY dist/apps/agora/app/browser/browser ${APP_DIR}/dist/apps/agora/app/browser/browser
+COPY dist/apps/agora/app/server ./server
+COPY dist/apps/agora/app/browser ./browser
 
 HEALTHCHECK --interval=2s --timeout=3s --retries=20 --start-period=5s \
   CMD curl --fail --silent "localhost:4200/health" | jq '.status' | grep UP || exit 1
@@ -21,4 +20,4 @@ EXPOSE 4200
 
 ENTRYPOINT ["/docker-entrypoint.sh"]
 
-CMD ["node", "main.js"]
+CMD ["node", "server/server.mjs"]