--- a/apps/agora/app/docker-entrypoint.d/10-envsubst-on-app-config-template.sh
+++ b/apps/agora/app/docker-entrypoint.d/10-envsubst-on-app-config-template.sh
@@ -1,4 +1,12 @@
 #!/usr/bin/env sh
 
-cd "${APP_DIR}/dist/apps/agora/app/browser/browser/config"
-envsubst < config.json.template > config.json
+# Generate 'config.json' from 'config.json.template' using environment variables.
+cd "${APP_DIR}/browser/config"
+envsubst < config.json.template | jq '
+  with_entries(
+    if .value == "true" then .value = true
+    elif .value == "false" then .value = false
+    else .
+    end
+  )
+' > config.json