--- a/apps/agora/app/docker-entrypoint.d/10-envsubst-on-app-config-template.sh
+++ b/apps/agora/app/docker-entrypoint.d/10-envsubst-on-app-config-template.sh
@@ -1,12 +1,4 @@
 #!/usr/bin/env sh
 
-# Generate 'config.json' from 'config.json.template' using environment variables.
 cd "${APP_DIR}/browser/config"
-envsubst < config.json.template | jq '
-  with_entries(
-    if .value == "true" then .value = true
-    elif .value == "false" then .value = false
-    else .
-    end
-  )
-' > config.json
+envsubst < config.json.template > config.json