--- a/apps/agora/app/src/app/app.config.ts
+++ b/apps/agora/app/src/app/app.config.ts
@@ -29,19 +29,9 @@ import { CustomUrlSerializer } from './app.custom-uri-serializer';
 import { routes } from './app.routes';
 import { provideClientHydration } from '@angular/platform-browser';
 
-// This index is used to remove the corresponding provider in app.config.server.ts.
-// TODO: This index could be out of sync if we are not careful. Find a more elegant way.
-export const APP_BASE_URL_PROVIDER_INDEX = 1;
-
 export const appConfig: ApplicationConfig = {
   providers: [
     { provide: APP_ID, useValue: 'agora-app' },
-    {
-      // This provider must be specified at the index defined by APP_BASE_URL_PROVIDER_INDEX.
-      provide: 'APP_BASE_URL',
-      useFactory: () => '.',
-      deps: [],
-    },
     {
       provide: SYNAPSE_API_CLIENT_BASE_PATH,
       useFactory: () => 'https://repo-prod.prod.sagebase.org',