--- a/apps/agora/app/src/styles.scss
+++ b/apps/agora/app/src/styles.scss
@@ -1,2 +1,2 @@
 /* You can add global styles to this file, and also import other style files */
-@use 'libs/agora/styles/src/index';
+@use 'agora/styles/src/index';