--- a/apps/model-ad/app/docker-entrypoint.d/10-envsubst-on-app-config-template.sh
+++ b/apps/model-ad/app/docker-entrypoint.d/10-envsubst-on-app-config-template.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env sh
 
-cd "${APP_DIR}/dist/apps/model-ad/app/browser/browser/config"
+cd "${APP_DIR}/browser/config"
 envsubst < config.json.template > config.json