--- a/apps/model-ad/app/server.ts
+++ b/apps/model-ad/app/server.ts
@@ -1,43 +1,41 @@
-import 'zone.js/node';
-
 import { APP_BASE_HREF } from '@angular/common';
 import { CommonEngine } from '@angular/ssr/node';
-import * as express from 'express';
-import { existsSync } from 'node:fs';
-import path, { join } from 'node:path';
+import express from 'express';
+import { dirname, join, resolve } from 'node:path';
+import { fileURLToPath } from 'node:url';
 import bootstrap from './src/main.server';
 
 // The Express app is exported so that it can be used by serverless Functions.
 export function app(): express.Express {
   const server = express();
-  const distFolder = join(process.cwd(), 'dist/apps/model-ad/app/browser/browser');
-  const indexHtml = existsSync(join(distFolder, 'index.original.html'))
-    ? join(distFolder, 'index.original.html')
-    : join(distFolder, 'index.html');
+  const serverDistFolder = dirname(fileURLToPath(import.meta.url));
+  const browserDistFolder = resolve(serverDistFolder, '../browser');
+  const indexHtml = join(serverDistFolder, 'index.server.html');
 
   const commonEngine = new CommonEngine();
 
   server.set('view engine', 'html');
-  server.set('views', distFolder);
+  server.set('views', browserDistFolder);
 
   // Example Express Rest API endpoints
   // server.get('/api/**', (req, res) => { });
   // Serve static files from /browser
   server.get(
-    '*.*',
-    express.static(distFolder, {
+    '**',
+    express.static(browserDistFolder, {
       maxAge: '1y',
+      index: 'index.html',
     }),
   );
 
-  // Health endpoint used by the container
+  // Health endpoint used by the Docker container
   server.get('/health', (_req, res) => res.status(200).json({ status: 'UP' }));
 
   // Serve static OpenGraph images at /opengraph
-  server.use('/opengraph', express.static(path.join(__dirname, 'public/images/opengraph')));
+  server.use('/opengraph', express.static(join(serverDistFolder, '../browser/opengraph')));
 
   // All regular routes use the Angular engine
-  server.get('*', (req, res, next) => {
+  server.get('**', (req, res, next) => {
     const { protocol, originalUrl, baseUrl, headers } = req;
 
     commonEngine
@@ -45,17 +43,8 @@ export function app(): express.Express {
         bootstrap,
         documentFilePath: indexHtml,
         url: `${protocol}://${headers.host}${originalUrl}`,
-        publicPath: distFolder,
-        providers: [
-          { provide: APP_BASE_HREF, useValue: baseUrl },
-          // The base URL enables the app to load the app config file during server-side rendering.
-          {
-            provide: 'APP_BASE_URL',
-            // the format of ${host} is `host:port`
-            useFactory: () => `${protocol}://${headers.host}`,
-            deps: [],
-          },
-        ],
+        publicPath: browserDistFolder,
+        providers: [{ provide: APP_BASE_HREF, useValue: baseUrl }],
       })
       .then((html) => res.send(html))
       .catch((err) => next(err));
@@ -74,15 +63,4 @@ function run(): void {
   });
 }
 
-// Webpack will replace 'require' with '__webpack_require__'
-// '__non_webpack_require__' is a proxy to Node 'require'
-// The below code is to ensure that the server is run only when not requiring the bundle.
-/* eslint-disable camelcase,no-undef */
-declare const __non_webpack_require__: NodeRequire;
-const mainModule = __non_webpack_require__.main;
-const moduleFilename = (mainModule && mainModule.filename) || '';
-if (moduleFilename === __filename || moduleFilename.includes('iisnode')) {
-  run();
-}
-
-export default bootstrap;
+run();