--- a/apps/model-ad/app/src/styles.scss
+++ b/apps/model-ad/app/src/styles.scss
@@ -1,3 +1,3 @@
 /* You can add global styles to this file, and also import other style files */
-@use 'styles/src/index';
 @use 'explorers/styles/src/index' as explorers;
+@use 'styles/src/index';