--- a/libs/agora/gene-comparison-tool/src/lib/components/gene-comparison-tool-details-panel/gene-comparison-tool-details-panel.component.scss
+++ b/libs/agora/gene-comparison-tool/src/lib/components/gene-comparison-tool-details-panel/gene-comparison-tool-details-panel.component.scss
@@ -1,7 +1,7 @@
 /* stylelint-disable no-descending-specificity */
 
-@use 'libs/agora/styles/src/lib/variables';
-@use 'libs/agora/styles/src/lib/mixins';
+@use 'agora/styles/src/lib/variables';
+@use 'agora/styles/src/lib/mixins';
 
 .gct-details-panel {
   .p-popover {