--- a/libs/agora/gene-comparison-tool/src/lib/components/gene-comparison-tool-filter-list-item/gene-comparison-tool-filter-list-item.component.scss
+++ b/libs/agora/gene-comparison-tool/src/lib/components/gene-comparison-tool-filter-list-item/gene-comparison-tool-filter-list-item.component.scss
@@ -1,6 +1,6 @@
 @use 'sass:map';
-@import 'libs/agora/styles/src/lib/variables';
-@import 'libs/agora/styles/src/lib/mixins';
+@use 'agora/styles/src/lib/variables';
+@use 'agora/styles/src/lib/mixins';
 
 .gct-filter-list-item {
   display: flex;