--- a/libs/agora/gene-comparison-tool/src/lib/components/gene-comparison-tool-filter-list-item/gene-comparison-tool-filter-list-item.component.scss
+++ b/libs/agora/gene-comparison-tool/src/lib/components/gene-comparison-tool-filter-list-item/gene-comparison-tool-filter-list-item.component.scss
@@ -5,7 +5,7 @@
 .gct-filter-list-item {
   display: flex;
   padding: 8px 15px;
-  background-color: rgba(map.get($main-colors, 'action-primary'), 0.2);
+  background-color: rgba(map.get(variables.$main-colors, 'action-primary'), 0.2);
   border-radius: 40px;
   margin-left: 8px;
   margin-bottom: 8px;
@@ -33,7 +33,7 @@
     color: var(--color-action-primary);
 
     svg path {
-      transition: $transition-duration;
+      transition: variables.$transition-duration;
     }
   }
 }