--- a/libs/agora/gene-comparison-tool/src/lib/components/gene-comparison-tool-filter-list/gene-comparison-tool-filter-list.component.scss
+++ b/libs/agora/gene-comparison-tool/src/lib/components/gene-comparison-tool-filter-list/gene-comparison-tool-filter-list.component.scss
@@ -6,7 +6,7 @@
   white-space: nowrap;
 
   button {
-    @include reset-button;
+    @include mixins.reset-button;
   }
 
   .gct-filter-list-inner {
@@ -25,12 +25,12 @@
     font-weight: 700;
     text-transform: uppercase;
     cursor: pointer;
-    transition: $transition-duration;
+    transition: variables.$transition-duration;
 
     &:hover {
       color: var(--color-action-primary);
 
-      @include link-hover;
+      @include mixins.link-hover;
     }
 
     &:hover agora-svg-icon {