--- a/libs/agora/gene-comparison-tool/src/lib/components/gene-comparison-tool-legend-panel/gene-comparison-tool-legend-panel.component.scss
+++ b/libs/agora/gene-comparison-tool/src/lib/components/gene-comparison-tool-legend-panel/gene-comparison-tool-legend-panel.component.scss
@@ -1,5 +1,5 @@
-@import 'libs/agora/styles/src/lib/variables';
-@import 'libs/agora/styles/src/lib/mixins';
+@use 'agora/styles/src/lib/variables';
+@use 'agora/styles/src/lib/mixins';
 
 .gct-legend-panel {
   color: var(--color-text);