--- a/libs/agora/gene-comparison-tool/src/lib/components/gene-comparison-tool-pinned-genes-modal/gene-comparison-tool-pinned-genes-modal.component.scss
+++ b/libs/agora/gene-comparison-tool/src/lib/components/gene-comparison-tool-pinned-genes-modal/gene-comparison-tool-pinned-genes-modal.component.scss
@@ -1,5 +1,5 @@
-@import 'libs/agora/styles/src/lib/variables';
-@import 'libs/agora/styles/src/lib/mixins';
+@use 'agora/styles/src/lib/variables';
+@use 'agora/styles/src/lib/mixins';
 
 .gct-pinned-genes-conversion-modal {
   max-height: 95% !important;