--- a/libs/agora/gene-comparison-tool/src/lib/components/gene-comparison-tool-score-panel/gene-comparison-tool-score-panel.component.scss
+++ b/libs/agora/gene-comparison-tool/src/lib/components/gene-comparison-tool-score-panel/gene-comparison-tool-score-panel.component.scss
@@ -1,5 +1,5 @@
-@use 'libs/agora/styles/src/lib/variables';
-@use 'libs/agora/styles/src/lib/mixins';
+@use 'agora/styles/src/lib/variables';
+@use 'agora/styles/src/lib/mixins';
 
 .gct-score-panel {
   .p-popover {