--- a/libs/agora/genes/src/lib/components/download-dom-image/download-dom-image.component.scss
+++ b/libs/agora/genes/src/lib/components/download-dom-image/download-dom-image.component.scss
@@ -1,5 +1,5 @@
-@use 'libs/agora/styles/src/lib/variables';
-@use 'libs/agora/styles/src/lib/mixins';
+@use 'agora/styles/src/lib/variables';
+@use 'agora/styles/src/lib/mixins';
 
 .download-dom-image-icon {
   transform: translateY(6px);