--- a/libs/agora/genes/src/lib/components/gene-details/gene-details.component.scss
+++ b/libs/agora/genes/src/lib/components/gene-details/gene-details.component.scss
@@ -1,8 +1,8 @@
 /* stylelint-disable plugin/no-unsupported-browser-features */
 /* stylelint-disable no-descending-specificity */
 
-@import 'libs/agora/styles/src/lib/variables';
-@import 'libs/agora/styles/src/lib/mixins';
+@use 'agora/styles/src/lib/variables';
+@use 'agora/styles/src/lib/mixins';
 
 .gene-details-nav {
   position: relative;