--- a/libs/agora/genes/src/lib/components/gene-hero/gene-hero.component.scss
+++ b/libs/agora/genes/src/lib/components/gene-hero/gene-hero.component.scss
@@ -1,5 +1,5 @@
-@import 'libs/agora/styles/src/lib/variables';
-@import 'libs/agora/styles/src/lib/mixins';
+@use 'agora/styles/src/lib/variables';
+@use 'agora/styles/src/lib/mixins';
 
 hr {
   border-color: #fff;