--- a/libs/agora/genes/src/lib/components/gene-nominations/gene-nominations.component.scss
+++ b/libs/agora/genes/src/lib/components/gene-nominations/gene-nominations.component.scss
@@ -7,7 +7,7 @@
   }
 
   p-button {
-    @include respond-to('ex-small') {
+    @include mixins.respond-to('ex-small') {
       width: 70%;
     }