--- a/libs/agora/genes/src/lib/components/popover-link/popover-link.component.scss
+++ b/libs/agora/genes/src/lib/components/popover-link/popover-link.component.scss
@@ -1,4 +1,4 @@
-@import 'libs/agora/styles/src/lib/mixins';
+@use 'agora/styles/src/lib/mixins';
 
 .popover-link-container {
   button {