--- a/libs/explorers/constants/src/index.ts
+++ b/libs/explorers/constants/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/injection-tokens';