… apps/agora/app/project.json
… apps/agora/app/tsconfig.app.json
… apps/model-ad/app/.env.example
… apps/model-ad/app/project.json
… apps/model-ad/app/tsconfig.app.json
… libs/agora/genes/src/lib/components/gene-evidence-metabolomics/gene-evidence-metabolomics.component.html
… libs/agora/genes/src/lib/components/gene-evidence-proteomics/gene-evidence-proteomics.component.html
… libs/agora/genes/src/lib/components/gene-evidence-rna/gene-evidence-rna.component.html
… libs/agora/genes/src/lib/components/gene-evidence-rna/gene-evidence-rna.component.ts
… libs/agora/genes/src/lib/components/gene-network/gene-network.component.html
… libs/agora/genes/src/lib/components/gene-resources/gene-resources.component.scss
… libs/agora/genes/src/lib/components/gene-search/gene-search.component.scss
… libs/agora/genes/src/lib/components/gene-search/gene-search.component.ts
… libs/agora/genes/src/lib/components/gene-soe-charts/gene-soe-charts.component.html
… libs/agora/genes/src/lib/components/gene-soe-charts/gene-soe-charts.component.scss
… libs/agora/genes/src/lib/components/gene-soe-list/gene-soe-list.component.scss
… libs/agora/genes/src/lib/components/gene-table/gene-table.component.ts
… libs/agora/home/src/lib/home.component.scss
… libs/agora/news/src/lib/news.component.scss
… libs/agora/nominated-targets/src/lib/nomination-form/nomination-form.component.scss
… libs/agora/not-found/src/lib/not-found.component.scss
… libs/agora/services/src/lib/helper.service.ts
… libs/agora/services/src/lib/svg-icon.service.ts
… libs/agora/shared/src/index.ts
… libs/agora/shared/src/lib/loading-icon/loading-icon.component.scss
… libs/agora/shared/src/lib/modal-link/modal-link.component.scss
… libs/agora/shared/src/lib/svg-icon/svg-icon.component.ts
… libs/agora/shared/src/lib/wiki/wiki.component.scss
… libs/agora/shared/src/lib/wiki/wiki.component.ts
… libs/agora/storybook/.storybook/main.ts