## Description

Migrates the Agora frontend to Vite for faster builds and modern DX. Based on the Vite setup used in OpenChallenges for consistency.

## Boosting Web App Build Performance and DX with Vite

### Before

Bundling the app with Webpack and serving it with the Webpack development server (`@angular-devkit/build-angular:browser` and `@angular-devkit/build-angular:dev-server`  Nx executors).

Build time: 56 seconds

```console
$ nx serve agora-app

...
Build at: 2025-06-27T02:30:52.931Z - Hash: cca5faf3de74c05f - Time: 55997ms
```

### After

Bundling the app with Vite and serving it with the Vite development server (`@nx/angular:application` and `@nx/angular:dev-server` Nx executors).

Build time: 27 seconds

```console
$ nx serve agora-app

...
Application bundle generation complete. [26.692 seconds]
```

## Related Issue

- [AG-1791](https://sagebionetworks.jira.com/browse/AG-1791)

## Changelog

- Update the `project.json` file of `agora-app` to use the Vite bundler.
- Replace `@import 'libs/agora/...'` by `@use 'agora/...'`.
- Update the `server.ts` of Agora and related files.
- Update the Dockerfile of `agora-app` to account for the new bundle.

## Preview

```bash
$ nx build agora-app
 NX   agora-app:serve:development                                                                                                                                                                   ↑
                                                                                                                                                                                                    █
  > nx run agora-app:serve:development                                                                                                                                                              █
                                                                                                                                                                                                    █
  Browser bundles                                                                                                                                                                                   █
  Initial chunk files  | Names            |  Raw size                                                                                                                                               █
  main.js              | main             |   1.73 MB |                                                                                                                                             █
  chunk-MH4NHW2M.js    | -                | 618.78 kB |                                                                                                                                             █
  chunk-4GS3BY2Q.js    | -                |  59.69 kB |                                                                                                                                             █
  chunk-RTBUHXBQ.js    | -                |  49.53 kB |                                                                                                                                             █
  styles.css           | styles           |  46.71 kB |                                                                                                                                             █
  chunk-5QLXPGVM.js    | -                |  41.55 kB |                                                                                                                                             █
  chunk-OCVIQ5QE.js    | -                |  18.21 kB |                                                                                                                                             █
  chunk-EVSPGG2W.js    | -                |   1.64 kB |                                                                                                                                             █
  polyfills.js         | polyfills        |  95 bytes |                                                                                                                                             █
                                                                                                                                                                                                    █
                       | Initial total    |   2.56 MB                                                                                                                                               █
                                                                                                                                                                                                    █
  Lazy chunk files     | Names            |  Raw size                                                                                                                                               █
  chunk-2P2Z2CGI.js    | index            | 381.40 kB |                                                                                                                                             █
  chunk-W7NMD35I.js    | index            |  39.56 kB |                                                                                                                                             █
  chunk-BQL7KD4P.js    | index            |  31.53 kB |                                                                                                                                             █
  chunk-EDLO5EAL.js    | index            |  26.57 kB |                                                                                                                                             █
  chunk-TPNBLG3C.js    | index            |  13.80 kB |                                                                                                                                             █
  chunk-TP32UKHA.js    | index            |  13.43 kB |                                                                                                                                             █
  chunk-HBDBNTL6.js    | index            |   3.06 kB |                                                                                                                                             █
  chunk-JMSPQFIB.js    | index            |   1.38 kB |                                                                                                                                             █
                                                                                                                                                                                                    █
                                                                                                                                                                                                    █
  Server bundles                                                                                                                                                                                    █
  Initial chunk files  | Names            |  Raw size                                                                                                                                               █
  chunk-WIPFPF44.mjs   | -                |   1.73 MB |                                                                                                                                             █
  chunk-27DLVDJL.mjs   | -                | 618.82 kB |                                                                                                                                             █
  chunk-PYS4B3JR.mjs   | -                |  59.72 kB |                                                                                                                                             █
  chunk-VJSGNZ45.mjs   | -                |  49.50 kB |                                                                                                                                             █
  chunk-NDYK663T.mjs   | -                |  41.59 kB |                                                                                                                                             █
  chunk-YB5XBK22.mjs   | -                |  18.30 kB |                                                                                                                                             ║
  server.mjs           | server           |   1.79 kB |                                                                                                                                             ║
  chunk-3ODA65P6.mjs   | -                |   1.67 kB |                                                                                                                                             ║
  main.server.mjs      | main.server      | 986 bytes |                                                                                                                                             ║
  polyfills.server.mjs | polyfills.server | 266 bytes |                                                                                                                                             ║
                                                                                                                                                                                                    ║
  Lazy chunk files     | Names            |  Raw size                                                                                                                                               ║
  chunk-EAPOS64W.mjs   | index            | 381.44 kB |                                                                                                                                             ║
  chunk-QFA4QBNI.mjs   | index            |  39.59 kB |                                                                                                                                             ║
  chunk-VUTHXAW3.mjs   | index            |  31.57 kB |                                                                                                                                             ║
  chunk-KQJRR46C.mjs   | index            |  26.61 kB |                                                                                                                                             ║
  chunk-ZEPHZUX4.mjs   | index            |  13.84 kB |                                                                                                                                             ║
  chunk-6ZIPNPQJ.mjs   | index            |  13.46 kB |                                                                                                                                             ║
  chunk-TMTE3NNH.mjs   | index            |   3.10 kB |                                                                                                                                             ║
  chunk-GB5U7JJW.mjs   | index            |   1.42 kB |                                                                                                                                             ║
                                                                                                                                                                                                    ║
  Application bundle generation complete. [26.692 seconds]

  Watch mode enabled. Watching for file changes...                                                                                                                                                  █
  NOTE: Raw file sizes do not reflect development server per-request transformations.                                                                                                               █
    ➜  Local:   http://localhost:4200/                                                                                                                                                              █
    ➜  press h + enter to show help
```

> [!NOTE]
> There are a couple of warning displayed about the library `angular-sass` that is now deprecated.

> [!NOTE]
> The error "Error loading commit SHA: ReferenceError: ErrorEvent is not defined" is shown but is unrelated to the migration to Vite as it is being observed on `main`. This issue is tracked on Jira [here](https://sagebionetworks.jira.com/browse/MG-269).

Preview:

![image](https://github.com/user-attachments/assets/61911269-e3a9-4d25-b0de-6762332ceddb)

### Live Reloading

The development server comes with live reloading. When saving a file, a message in the VS Code terminal shows that a changes has been detected and that Vite is regenerating the bundle. As soon as the logs shown that the bundle has been regenerated, the app opened in the browser (here Chrome) reload automatically virtually instantly, even thought I'm developing remotely on an EC2 instance.

```console
  Browser bundles                                                                                                                                                                                   ║
  Initial chunk files | Names       |  Raw size                                                                                                                                                     ║
  main.js             | main        |   1.73 MB |                                                                                                                                                   ║
                                                                                                                                                                                                    ║
                                                                                                                                                                                                    ║
  Server bundles                                                                                                                                                                                    █
  Initial chunk files | Names       |  Raw size                                                                                                                                                     █
  chunk-GTET7F22.mjs  | -           |   1.73 MB |                                                                                                                                                   █
  server.mjs          | server      |   1.79 kB |                                                                                                                                                   █
  main.server.mjs     | main.server | 986 bytes |                                                                                                                                                   █
                                                                                                                                                                                                    █
  Application bundle generation complete. [7.013 seconds]                                                                                                                                           █
                                                                                                                                                                                                    █
  Component update sent to client(s).
```

### Start with Docker Compose

```bash
agora-build-images
nx serve-detach agora-apex
```




[AG-1791]: https://sagebionetworks.jira.com/browse/AG-1791?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ