@tschaffter 2025-06-25
## File 'apps/agora/app/server.ts' is missing from the TypeScript compilation.

```console
$ nx build agora-app

...
✘ [ERROR] File 'apps/agora/app/server.ts' is missing from the TypeScript compilation. [plugin angular-compiler]

  Ensure the file is part of the TypeScript program via the 'files' or 'include' property.
```