@tschaffter 2025-06-25
## Fix Watchpack Error (watcher): Error: EMFILE: too many open files, watch '<a directory>'

Running the dev server of the Agora web app shows the following errors:

```console
$ nx serve agora-app

...
Watchpack Error (watcher): Error: EMFILE: too many open files, watch '/workspaces/sage-monorepo'
Watchpack Error (watcher): Error: EMFILE: too many open files, watch '/workspaces'
Watchpack Error (watcher): Error: EMFILE: too many open files, watch '/workspaces/sage-monorepo/libs/agora/ui/src/lib/components'
Watchpack Error (watcher): Error: EMFILE: too many open files, watch '/workspaces/sage-monorepo/libs/agora/ui/src/lib'
Watchpack Error (watcher): Error: EMFILE: too many open files, watch '/workspaces/sage-monorepo/libs/synapse/api-client-angular/src/lib'
Watchpack Error (watcher): Error: EMFILE: too many open files, watch '/workspaces/sage-monorepo/libs/synapse/api-client-angular/src'
Watchpack Error (watcher): Error: EMFILE: too many open files, watch '/workspaces/sage-monorepo/libs/synapse/api-client-angular'
Watchpack Error (watcher): Error: EMFILE: too many open files, watch '/workspaces/sage-monorepo/libs/synapse'
Watchpack Error (watcher): Error: EMFILE: too many open files, watch '/workspaces/sage-monorepo/apps/agora/app'
Watchpack Error (watcher): Error: EMFILE: too many open files, watch '/workspaces/sage-monorepo/apps/agora'
```

This issue has happened to me recently and more frequently. This issue is tracked in this [Jira ticket](https://sagebionetworks.jira.com/browse/SMR-203).

Meanwhile, the app can be started without watching for file changes:

```bash
nx serve agora-app --watch false
```