@tschaffter 2025-06-30
```console
✘ [ERROR] Can't find stylesheet to import.
  ╷
1 │ @use 'libs/explorers/styles/src/lib/variables';
  │ ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  ╵
  libs/explorers/util/src/lib/loading-icon/loading-icon.component.scss 1:1  root stylesheet [plugin angular-sass]
```

The shared `explorers` components need to be updated, which will require migrating Model-AD to Vite. @hallieswan, would you be able to update Model-AD using the same approach implemented in this PR?