@hallieswan 2025-06-30
@tschaffter I also noticed that Agora plots are not rendering correctly on this branch. To reproduce: 
1. Build and run agora: `agora-build-images && agora-docker-start`
2. Navigate to RNA evidence page for a gene: `http://localhost:8000/genes/ENSG00000171862/evidence/rna`
3. Scroll down to Consistency of Change in Expression plot: 

plot on `main` | plot on `agora/vite`
:----:|:----:
<img width="2153" alt="agora-plot-main" src="https://github.com/user-attachments/assets/bb505efe-bf19-41bf-a140-c325a5d2cba0" /> | <img width="2147" alt="agora-plot-vite" src="https://github.com/user-attachments/assets/717a0c4a-0e55-441a-963b-48cdeba20144" />