@tschaffter 2025-07-01
@hallieswan The plot looks good on my end. I did rebase this PR so that may have solved the issue. The web console doesn't show any errors. Here is a screenshot of the page. Does it look alright?

`http://localhost:8000/genes/ENSG00000171862/evidence/rna`:

![image](https://github.com/user-attachments/assets/adec3c3c-0d27-498f-be85-a04d9955ce4a)