@hallieswan 2025-07-01
> @hallieswan The plot looks good on my end; actually, more than great if I'm being honest 🤩. I did rebase this PR so that may have solved the issue. The web console doesn't show any errors. Here is a screenshot of the page. Does it look alright?

@tschaffter That does look right! I tried rebuilding with the latest changes on this branch, but I'm still seeing the same problem with the plots (tested in Chrome, Safari, and Firefox). @sagely1 would you mind being our tiebreaker and testing the build in your environment? Hopefully it's just an issue on my end!