@sagely1 2025-07-01
I think I see the problem.  Navigating to the url directly is broken.  However, navigating to the PTEN gene via the header does work. http://localhost:8000/genes/ENSG00000171862/evidence/rna
If you load this url directly you will see the issue:
![image](https://github.com/user-attachments/assets/67ea0d29-99de-470a-ac25-71a3972ecb07)