@tschaffter 2025-07-03
@hallieswan `model-ad:serve` throws this error when accessing its home page:

```console
Watch mode enabled. Watching for file changes...
NOTE: Raw file sizes do not reflect development server per-request transformations.
  ➜  Local:   http://localhost:4200/
  ➜  press h + enter to show help
ERROR ReferenceError: getComputedStyle is not defined
    at new _HomeCardComponent (/workspaces/sage-monorepo/libs/explorers/ui/src/lib/components/home-card/home-card.component.ts:32:5)
    at Component (/workspaces/sage-monorepo/libs/explorers/ui/src/lib/components/home-card/home-card.component.ts:35:3)
    at getNodeInjectable (/workspaces/sage-monorepo/.angular/cache/20.0.3/model-ad-app/vite/deps_ssr/chunk-2O2DLPQO.js:1445:38)
    at instantiateAllDirectives (/workspaces/sage-monorepo/.angular/cache/20.0.3/model-ad-app/vite/deps_ssr/chunk-2O2DLPQO.js:4445:23)
    at createDirectivesInstances (/workspaces/sage-monorepo/.angular/cache/20.0.3/model-ad-app/vite/deps_ssr/chunk-2O2DLPQO.js:4298:3)
    at ɵɵelementStart (/workspaces/sage-monorepo/.angular/cache/20.0.3/model-ad-app/vite/deps_ssr/chunk-2O2DLPQO.js:14182:5)
    at Module.ɵɵelement (/workspaces/sage-monorepo/.angular/cache/20.0.3/model-ad-app/vite/deps_ssr/chunk-2O2DLPQO.js:14223:3)
    at HomeComponent_Template (/workspaces/sage-monorepo/libs/model-ad/home/src/lib/home.component.html:21:9)
    at executeTemplate (/workspaces/sage-monorepo/.angular/cache/20.0.3/model-ad-app/vite/deps_ssr/chunk-2O2DLPQO.js:4290:5)
    at renderView (/workspaces/sage-monorepo/.angular/cache/20.0.3/model-ad-app/vite/deps_ssr/chunk-2O2DLPQO.js:4666:7)
```